feat(docker): Added 'run mode' for docker plugin + some styling

+++ plugins/docker/oper-kill.sh
@@ -1,18 +1,16 @@
 #!/bin/bash
 
 : "${taito_cli_path:?}"
+: "${taito_plugin_path:?}"
 : "${taito_project:?}"
 
-pod="${1:?Pod name not given}"
-
-if [[ ${pod} != *"-"* ]]; then
-  pod="${taito_project}-${pod}"
-fi
-
 echo
-echo "### docker - oper-kill: Killing in the name of ${pod} ###"
+echo "### docker - oper-kill: Killing ###"
 
-"${taito_cli_path}/util/execute-on-host-fg.sh" "docker kill ${pod}" && \
+# shellcheck disable=SC1090
+. "${taito_plugin_path}/util/determine-pod.sh" "${@}" && \
+
+"${taito_cli_path}/util/execute-on-host-fg.sh" "docker kill ${pod:?}" && \
 
 # Call next command on command chain
 "${taito_cli_path}/util/call-next.sh" "${@}"
